refactor(form): colocate reference item together with reference field and input
