+++ packages/sanity/src/core/form/inputs/ReferenceInput/ReferenceField.tsx
@@ -22,12 +22,12 @@ import {
 } from '@sanity/icons'
 import {ObjectFieldProps, RenderPreviewCallback} from '../../types'
 import {FormField} from '../../components'
-import {useReferenceInput} from '../arrays/ArrayOfObjectsInput/_reference/useReferenceInput'
 import {useScrollIntoViewOnFocusWithin} from '../../hooks/useScrollIntoViewOnFocusWithin'
 import {useDidUpdate} from '../../hooks/useDidUpdate'
 import {set, unset} from '../../patch'
-import {ReferencePreviewCard} from '../arrays/ArrayOfObjectsInput/List/ReferenceItem'
 import {AlertStrip} from '../../components/AlertStrip'
+import {ReferencePreviewCard} from './ReferenceItem'
+import {useReferenceInput} from './useReferenceInput'
 import {useReferenceInfo} from './useReferenceInfo'
 import {PreviewReferenceValue} from './PreviewReferenceValue'
