+++ packages/sanity/src/core/form/inputs/ReferenceInput/ReferenceItem.tsx
@@ -23,22 +23,22 @@ import {
   TrashIcon,
 } from '@sanity/icons'
 import styled from 'styled-components'
-import {ObjectItem, ObjectItemProps} from '../../../../types'
-import {useScrollIntoViewOnFocusWithin} from '../../../../hooks/useScrollIntoViewOnFocusWithin'
-import {useDidUpdate} from '../../../../hooks/useDidUpdate'
-import {randomKey} from '../../../../utils/randomKey'
-import {FormFieldSet, FormFieldValidationStatus} from '../../../../components/formField'
-import {FieldPresence} from '../../../../../presence'
+import {ObjectItem, ObjectItemProps} from '../../types'
+import {useScrollIntoViewOnFocusWithin} from '../../hooks/useScrollIntoViewOnFocusWithin'
+import {useDidUpdate} from '../../hooks/useDidUpdate'
+import {randomKey} from '../../utils/randomKey'
+import {FormFieldSet, FormFieldValidationStatus} from '../../components/formField'
+import {FieldPresence} from '../../../presence'
 
-import {ChangeIndicator} from '../../../../../changeIndicators'
-import {RowLayout} from '../../layouts/RowLayout'
-import {PreviewReferenceValue} from '../../../ReferenceInput/PreviewReferenceValue'
-import {useReferenceInfo} from '../../../ReferenceInput/useReferenceInfo'
-import {AlertStrip} from '../../../../components/AlertStrip'
-import {set, unset} from '../../../../patch'
-import {createProtoArrayValue} from '../createProtoArrayValue'
-import {InsertMenu} from '../InsertMenu'
-import {useReferenceInput} from '../_reference/useReferenceInput'
+import {ChangeIndicator} from '../../../changeIndicators'
+import {RowLayout} from '../arrays/layouts/RowLayout'
+import {AlertStrip} from '../../components/AlertStrip'
+import {set, unset} from '../../patch'
+import {createProtoArrayValue} from '../arrays/ArrayOfObjectsInput/createProtoArrayValue'
+import {InsertMenu} from '../arrays/ArrayOfObjectsInput/InsertMenu'
+import {useReferenceInfo} from './useReferenceInfo'
+import {PreviewReferenceValue} from './PreviewReferenceValue'
+import {useReferenceInput} from './useReferenceInput'
 
 export interface ReferenceItemValue extends Omit<ObjectItem, '_type'>, Omit<Reference, '_key'> {}
